+++ internal/model/size_model.go
@@ -1,20 +1,20 @@
 package model
 
 type SizeResponse struct {
-	ID        uint   `json:"id,omitempty"`
-	Name      string `json:"name,omitempty"`
-	SellPrice uint   `json:"sell_price,omitempty"`
-	BuyPrice  uint   `json:"buy_price,omitempty"`
-	CreatedAt int64  `json:"created_at,omitempty"`
-	UpdatedAt int64  `json:"updated_at,omitempty"`
+	ID        uint    `json:"id,omitempty"`
+	Name      string  `json:"name,omitempty"`
+	SellPrice float64 `json:"sell_price,omitempty"`
+	BuyPrice  float64 `json:"buy_price,omitempty"`
+	CreatedAt int64   `json:"created_at,omitempty"`
+	UpdatedAt int64   `json:"updated_at,omitempty"`
 	// Product   ProductResponse `json:"product"`
 }
 
 type CreateSizeRequest struct {
-	ProductSKU string `json:"-" validate:"required,max=100"`
-	Name       string `json:"name" validate:"required,max=100"`
-	SellPrice  uint   `json:"sell_price,omitempty" validate:"required"`
-	BuyPrice   uint   `json:"buy_price,omitempty" validate:"required"`
+	ProductSKU string  `json:"-" validate:"required,max=100"`
+	Name       string  `json:"name" validate:"required,max=100"`
+	SellPrice  float64 `json:"sell_price,omitempty" validate:"required"`
+	BuyPrice   float64 `json:"buy_price,omitempty" validate:"required"`
 }
 
 type SearchSizeRequest struct {
@@ -30,11 +30,11 @@ type GetSizeRequest struct {
 }
 
 type UpdateSizeRequest struct {
-	ProductSKU string `json:"-" validate:"required,max=100"`
-	ID         uint   `json:"-" id:"required,max=100"`
-	Name       string `json:"name,omitempty" validate:"max=100"`
-	SellPrice  uint   `json:"sell_price,omitempty"`
-	BuyPrice   uint   `json:"buy_price,omitempty"`
+	ProductSKU string  `json:"-" validate:"required,max=100"`
+	ID         uint    `json:"-" id:"required,max=100"`
+	Name       string  `json:"name,omitempty" validate:"max=100"`
+	SellPrice  float64 `json:"sell_price,omitempty"`
+	BuyPrice   float64 `json:"buy_price,omitempty"`
 }
 
 type DeleteSizeRequest struct {
